Implement bitwise ops.
I'm not sure how complete this is, because benchmarks/plush_parser.pls
didn't have several important operators that plush/parser.pls already
had, and I did not add the bitwise operators to the file in benchmarks/
. Other than that, C++ and Plush files have been added to, and the only
change that may be an issue is line 1119 (previously 1113) of
plush/parser.pls, where it was necessary to reduce the amount of
precedence change. The case that precipitated this was that `||` would
be parsed as `|` followed by `|` if that operator was in the RHS of
another `||` call, as with `a || b || c`. Allowing the same precedence
level (which is always higher than 0, because no operators have 0
precedence) means `||` can be used with other `||` operations without
issue. This may have surfaced anyway if an operator like `|>` was
introduced and could be chained, since that might be parsed as `|` with
precedence 5 followed by `>` with precedence 9.