Setting elements in arrays works now.
The code is largely similar to how properties on objects are set, but
the subtle differences were a bit difficult at first. This is tested in
simple_exprs.pls, including setting an element of an array with =,
setting another element and comparing the two changed ones, and
assigning with *= . To ensure I didn't break anything in property
setting, I also added a test for setting an object property, combined
with adding two properties.

As an aside, building and testing ZetaVM seems faster, and I am guessing
the recent commits with performance in mind are to credit. It's a lot
more pleasant to work with now that compiles are faster!