Expand type checks with rt_setElem
Checks are also performed in rt_getElem. The changes from the
non-checking veersion from before to use rt_setElem appear to also need
a change to interp.cpp . This accounts for the result of the assignment
being a usable value, as opposed to just being discarded. There may be a
better way around this, but I am not especially familiar with stack
machines, and I'd need someone else to point out the better way.

This does not change anything relating to a setField instruction because
code that sets fields is already present and I don't want to interfere
with any changes that may need to happen in that code. I'm limiting my
changes to the scope of the current PR.