Update test, fix int method, add global RNG
The int method now uses the correct range, inclusive on both ends. An
RNG's state can be reseeded with the setState() method. The globalRNG
var exported by random is currently seeded the same every time; this
should probably change if there's a good way to feed a semi-random seed
to it. While the seed is the same, though, the result happens to choose
"Ace" from an array of values and "Spades" from an array of suits.

Double up or quit / double take or split / The Ace of Spades...