[Scheme] Import builtins into the top global object
This patch changes the runtime setup code to import
all the builtins into the top-level global object.
This slightly simplifies code generation and will
come in handy later when things are reworked to
allows application writers to override builtins.

The names of builtins are also currently mangled
because string literals can't be used as object
field names.  See:

  https://github.com/zetavm/zetavm/issues/62