Fix test running on Arch
I attempted to run the tests on Arch Linux, and due to the default Python version here being python3, "zetavm/espresso/e_codegen.py", line 30, failed as "has_key" does not exist in python3 (just use "in").

This should also fix it for other platforms where Python3 is the default. Those where Python2 is not set will get a much easier error to deal with anyway.